Moved type facility functions for deciding code generation type flags.
