+++ src/cc65/expr.h
@@ -51,6 +51,12 @@ typedef struct GenDesc {
 unsigned GlobalModeFlags (const ExprDesc* Expr);
 /* Return the addressing mode flags for the given expression */
 
+unsigned TypeOf (const Type* T);
+/* Get the code generator base type of the object */
+
+unsigned FuncTypeOf (const Type* T);
+/* Get the code generator flag for calling the function */
+
 void ExprWithCheck (void (*Func) (ExprDesc*), ExprDesc* Expr);
 /* Call an expression function with checks. */
