+++ src/cc65/function.c
@@ -42,6 +42,7 @@
 #include "asmlabel.h"
 #include "codegen.h"
 #include "error.h"
+#include "expr.h"
 #include "funcdesc.h"
 #include "global.h"
 #include "litpool.h"
